LayoutTests/platform/ios-simulator/fast/canvas/canvas-crash-expected.txt

CONSOLE MESSAGE: line 15: Total canvas memory use exceeds the maximum limit (448 MB).
CONSOLE MESSAGE: line 16: TypeError: null is not an object (evaluating 'ctx.putImageData')
